# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Highleads API",
        "description": "API highleads-backend",
        "contact": {
            "name": "Marco Aurélio",
            "email": "marcoascabral@gmail.com"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/v1/usuarios/{id}": {
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Alterar um registro",
                "description": "Alterar um registro por ID",
                "operationId": "f429f3e0b0128488cae87b8966cffadd",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID",
                        "required": true,
                        "schema": {
                            "type": "integer",
                            "format": "int64"
                        },
                        "example": "1"
                    }
                ],
                "requestBody": {
                    "description": "Dados do registro",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "nome",
                                    "email"
                                ],
                                "properties": {
                                    "nome": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "José da Silva"
                                    },
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Salvo com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao inserir."
                    },
                    "404": {
                        "description": "Nnão encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/v1/altera-senha": {
            "put": {
                "tags": [
                    "Usuários"
                ],
                "summary": "Altera senha do usuário",
                "description": "Alterar a senha do usuário cadastrado",
                "operationId": "545e0b1f8d409f396845d4f9a292e179",
                "requestBody": {
                    "description": "Dados do Usuario",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email",
                                    "codigo",
                                    "senha"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "email",
                                        "format": "text",
                                        "example": "mercedes68@example.org"
                                    },
                                    "codigo": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "JAVA"
                                    },
                                    "senha": {
                                        "type": "string",
                                        "format": "text",
                                        "example": "******"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Senha Alterada com sucesso.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "updated": {
                                            "type": "string",
                                            "example": "true"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Erro ao alterar a senha."
                    },
                    "404": {
                        "description": "Usuário não encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "mensagem": {
                                            "type": "string",
                                            "example": "Não encontrado"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        }
    }
}